semana8/requirements.txt
fastapi
uvicorn
pyjwt
pytest
httpx
slowapi
python-dotenv
pydantic-settings